Release jQuery as well as $ using .noConflict(true)
In order to use multiple versions of jquery on the same site you must use
jQuery.noConflict(true) to release both $ and jQuery.  Otherwise the jQuery
variable is still held by the last loaded version of jQuery which may not be the
correct version.

We store a reference to jQuery 1.7.2 in a local copy of the jQuery variable.
The referenced version of jQuery should revert back to the correct jQuery
version after the .noConflict call.