Refactor the enterButtonModal() function
Split the enterButtonModal function into two new functions incrementInvitationAcceptCount and openSurveyInNewWIndow to prepare for splitting the code paths between a pop-up window survey and a modal survey.  This will let us reuse some of the code for managing the accept analytics and give us a clean entry point for our divergent code path.