Remove caching of invitation modal
When the invitation code was copied over from GSA, the original caching
code was left in-place.  That code must be removed since the users must
be able to update the invitation text without having to add a cache
buster to the URL.

The code allowing for the {{accept}} {{reject}} tags was also not copied
over from GSA so we are adding that now.

This should fix the caching issues we have been seeing with the
foodsafety testing.