Fix caching issue with browser detection
Browser detection is done using the rubygem 'browser' and checking for
`browser.mobile? || browser.tablet?` inside erb tags in invitation.js.
This works as long as there is no caching.  If there is caching, then if
a desktop version is requested first, the desktop version of the
invitation.js is cached and served to all other requests without going
through the application server to re-evaluate the erb conditional to
include/exclude the device specific code.  This results in mobile and
table devices being served the invitaiton pop-up when they should not
be.

We now do the device check using the `jQuery.browser.mobile` function in
the invitation.js file.  This allows for the use of a single
invitation.js file to be cached regardless of whether a mobile device or
desktop is the first to hit the application server.

The code for the jQuery.browser.mobile function is taken from:

http://www.sitepoint.com/detect-mobile-devices-jquery/