Cancel partial submission timer on form submit
When the form is submitted but the partial submission timeout is still
running, it is possible to update the raw_submission submitted status
from true to false and cause the sweeper to treat it as an unsubmitted
survey.  This causes a duplicate survey response to be created.

Add a call to clearPartialTimeout to remove the partial submission
timeout on submission if all validations pass.