Scope javascript functions to a survey instance.
When embedding multiple surveys on the same page, the javascript
functions which control the survey behavior cannot tell which survey
instance they are operating on.

Add a reference to the link/button which triggers the event in order to
search up the DOM and find the parent form container for the survey.
Once a parent element is found, the subsequent survey control DOM
querries are scoped to the survey container using .find