Add content_id param for invitation.js
The modal sets the ARIA attributes for the main content body while the
modal is shown.  In order to know the correct div to apply the
aria-hidden label to, it must be passed in to the script tag as a
parameter.  If no parameter is passed in, it will default to
main-content.